requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.3
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.2
email-validator==2.0.0
Flask==2.2.5
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
passlib==1.7.4
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
SQLAlchemy==2.0.21
Werkzeug==2.2.3
WTForms==3.0.1